Added to_clipboard option, quotation changes, and order_by_attribute
Made a function for writing the mdx to the clipboard, so that can be quickly pasted into a TI.
Added quotation changes to some mdx outputs to align with TI code (' to ").
Added extra arg for filter by attribute, to allow for specifying the equality operator.
Added function and class to allow user to order by attribute values.